Refactor search functionality and component logic
This commit improves the maintainability and readability of the search functionality and related components by performing a series of refactorings.

- **Centralize Search Logic:** The duplicated search query parsing logic in `lib/search.ts` has been extracted into a single `parseQuery` function. Both `filterPostsBySearch` and `calculateScores` now use this shared function.

- **Separate Concerns in UI:**
  - A generic `useDebounce` hook has been created to handle value debouncing.
  - The `useSearchSuggestions` and `useSearch` hooks have been combined into a single `useSearchOverlay` hook, which now manages all state and logic for the search overlay.
  - The `SearchOverlay.tsx` component has been simplified to be a presentational component that uses the new hook.

- **Externalize Configuration:** Hardcoded values (magic numbers) related to search have been moved to a new centralized configuration file at `src/constants/searchConfig.ts`. The search API route and the search overlay component now import their configuration from this single source of truth.

- **Improve Documentation:** JSDoc comments have been added to `getNextPost` and `getPreviousPost` in `lib/post-filters.ts` to make the implicit requirement of a pre-sorted post array explicit.
